Refactor ModuleUtils::collectHeapTypes
Move the code into a new .cpp file because it is too complicated for a header
and update the API to make both the type indices and optimized sorting optional.
It will become more important to avoid unnecessary sorting once isorecursive
types have been implemented because they will make the sorting more complicated.
